kernel: Add CPU stress test driver for identifying CPU instability
This adds a driver that will automatically stress all CPU cores a short
time after init in order to identify CPU stability issues. After the
stress test begins, it will never stop, so this driver should only be
enabled when a stress test is desired.

This is useful for determining if your CPU hardware may be faulty.

Signed-off-by: Sultanxda <sultanxda@gmail.com>